% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_records.R
\name{delete_records}
\alias{delete_records}
\title{Delete records from a REDCap project}
\usage{
delete_records(conn, records)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{records}{Character vector of record IDs to delete}
}
\value{
An integer, the number of records successfully deleted
}
\description{
Delete records from a REDCap project
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

# delete all records associated with IDs "P004" and "P007"
delete_records(conn, records = c("P004", "P007"))
}

}
