% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_records.R
\name{import_records}
\alias{import_records}
\title{Import records into a REDCap project}
\usage{
import_records(conn, data, overwrite = "normal", return = "count")
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{data}{A data.frame containing record data to import into REDCap}

\item{overwrite}{Overwrite behaviour. Either "normal" to prevent missing
values from overwriting data, or "overwrite" to allow data to be
overwritten with missing values. Defaults to "normal".}

\item{return}{What to return. Use "count" to return a count of imported
records, "ids" to return a vector of the IDs that were imported, or
"nothing" to return nothing. Defaults to "count".}
}
\value{
Depends on argument \code{return}. Either a count of imported records, a vector of
record IDs, or nothing.
}
\description{
Import records into a REDCap project
}
