% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_arms.R
\name{meta_arms}
\alias{meta_arms}
\title{Export REDCap project arms}
\usage{
meta_arms(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 2 columns:
\itemize{
\item \code{arm_num}
\item \code{name}
}
}
\description{
Execute an "Export Arms" API request to fetch the "arms" (number and name)
associated with a REDCap project.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_arms(conn)
}

}
