% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dictionary.R
\name{meta_dictionary}
\alias{meta_dictionary}
\title{Fetch variable dictionary for a REDCap project}
\usage{
meta_dictionary(
  conn,
  forms = NULL,
  expand_checkbox = TRUE,
  cols_omit = c("section_header", "custom_alignment", "question_number",
    "matrix_group_name", "matrix_ranking")
)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{forms}{Character vector of forms (i.e. instruments) to include in the
return. Set to \code{NULL} (the default) to return dictionary entries for all
forms in the project.}

\item{expand_checkbox}{Logical indicating whether to expand checkbox
variables. Defaults to \code{TRUE}.

Unlike an "Export Records" API request (see \code{\link{fetch_records}}), which
returns 1 column for each checkbox \emph{option}, an "Export Metadata (Data
Dictionary)" request returns a single row for each \emph{field} — including
checkbox fields. Thus, the \code{field_name} and \code{field_label} entries for
checkbox variables in the data dictionary will never exactly match the
respective column names or values returned by \code{\link{fetch_records}}.

When \code{expand_checkbox} is \code{TRUE}, rows for checkbox fields are expanded to
1 row per checkbox \emph{option}, so that dictionary entries for \code{field_name},
\code{field_label}, and \code{choices} will always match the relevant entries
returned by \code{\link{fetch_records}}.}

\item{cols_omit}{Character vector of dictionary columns to omit from the
return for brevity. Set to \code{NULL} to return all columns.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing the project
dictionary. Note that some of the returned column names are shortened
versions of the original column names returned by the API:\tabular{ll}{
   \strong{Original} \tab \strong{Returned} \cr
   \code{select_choices_or_calculations} \tab \code{choices} \cr
   \code{text_validation_type_or_show_slider_number}  \tab \code{validation} \cr
   \code{text_validation_min} \tab \code{validation_min} \cr
   \code{text_validation_max} \tab \code{validation_max} \cr
}
}
\description{
Execute an "Export Metadata (Data Dictionary)" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing the project codebook
(field names, types, labels, choices, validation, etc.).
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_dictionary(conn)
}

}
