% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{meta_events}
\alias{meta_events}
\title{Fetch event names and labels for a REDCap project}
\usage{
meta_events(conn, on_error = "fail")
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{on_error}{How to handle errors returned by the API (e.g. events cannot
be exported for classic projects). Set to "fail" to halt execution and
return the API error message, or "null" to ignore the error and return
\code{NULL}. Defaults to "fail".}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 7 columns:
\itemize{
\item \code{event_name}
\item \code{arm_num}
\item \code{day_offset}
\item \code{offset_min}
\item \code{offset_max}
\item \code{unique_event_name}
\item \code{custom_event_label}
}
}
\description{
Execute an "Export Events" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing event names and
labels. Note that this request type is not available for 'classic projects',
from which event details cannot be exported.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_events(conn)
}

}
