% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_factors.R
\name{meta_factors}
\alias{meta_factors}
\title{Fetch field option values and labels for factor-type variables in a REDCap
project}
\usage{
meta_factors(
  conn,
  forms = NULL,
  expand_checkbox = TRUE,
  types = c("radio", "yesno", "dropdown", "checkbox")
)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{forms}{Character vector of forms (i.e. instruments) to include. Set to
\code{NULL} (the default) to return field options for all forms in the project.}

\item{expand_checkbox}{Logical indicating whether to expand checkbox
variables. Defaults to \code{TRUE}.

Unlike an "Export Records" API request (see \code{\link{fetch_records}}), which
returns 1 column for each checkbox \emph{option}, an "Export Metadata (Data
Dictionary)" request returns a single row for each \emph{field} — including
checkbox fields. Thus, the \code{field_name} and \code{field_label} entries for
checkbox variables in the data dictionary will never exactly match the
respective column names or values returned by \code{\link{fetch_records}}.

When \code{expand_checkbox} is \code{TRUE}, rows for checkbox fields are expanded to
1 row per checkbox \emph{option}, so that dictionary entries for \code{field_name},
\code{field_label}, and \code{choices} will always match the relevant entries
returned by \code{\link{fetch_records}}.}

\item{types}{Character vector of variable types to return field options for.
\cr Defaults to \code{c("radio", "yesno", "dropdown", "checkbox")}.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 6 columns:
\itemize{
\item field_name
\item form_name
\item field_type
\item field_label
\item value
\item label
}
}
\description{
Converts field options and labels from the metadata dictionary (see
\code{\link{meta_dictionary}}), which has 1 row per variable and field options in
compact string-form, to long form with 1 row per option. E.g.:

\strong{Dictionary version} (1 row per variable):\tabular{llll}{
   \code{field_name} \tab \code{form_name} \tab \code{field_type} \tab \code{choices} \cr
   consented \tab enrollment \tab radio \tab 0, No | 1, Yes \cr
}


\strong{Long format} (1 row per option):\tabular{lllll}{
   \code{field_name} \tab \code{form_name} \tab \code{field_type} \tab \code{value} \tab \code{label} \cr
   consented \tab enrollment \tab radio \tab 0 \tab No \cr
   consented \tab enrollment \tab radio \tab 1 \tab Yes \cr
}
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_factors(conn)
}

}
\seealso{
\code{\link{meta_dictionary}}
}
