% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_fields.R
\name{meta_fields}
\alias{meta_fields}
\title{Fetch exported field names for a REDCap project}
\usage{
meta_fields(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 3 columns:
\itemize{
\item \code{original_field_name}
\item \code{choice_value}
\item \code{export_field_name}
}
}
\description{
Execute an "Export List of Export Field Names" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing field (i.e. variable)
names as listed in the project codebook (column \code{original_field_name}) and
the corresponding exported variable name(s) (column \code{export_field_name}).

Original and exported field names will be identical except in the case of
checkbox-type variables. A given checkbox variable (e.g. "patient_status")
will have a single entry in the codebook (i.e. \code{field_name} =
"patient_status"), but will be exported as multiple variables — one for each
possible choice value.\tabular{lll}{
   \code{original_field_name} \tab \code{choice_value} \tab \code{export_field_name} \cr
   patient_status \tab 1 \tab patient_status___1 \cr
   patient_status \tab 2 \tab patient_status___2 \cr
   patient_status \tab 3 \tab patient_status___3 \cr
   patient_status \tab 88 \tab patient_status___88 \cr
}
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_fields(conn)
}

}
