% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_forms.R
\name{meta_forms}
\alias{meta_forms}
\title{Fetch instrument names and labels for a REDCap project}
\usage{
meta_forms(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 2 columns:
\itemize{
\item \code{instrument_name}
\item \code{instrument_label}
}
}
\description{
Execute an "Export Instrument (Data Entry Forms)" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing instrument names and
labels.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_forms(conn)
}

}
