% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_mapping.R
\name{meta_mapping}
\alias{meta_mapping}
\title{Fetch instrument-event mappings for a REDCap project}
\usage{
meta_mapping(conn, on_error = "fail")
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{on_error}{How to handle errors returned by the API (e.g. events cannot
be exported for classic projects). Set to "fail" to halt execution and
return the API error message, or "null" to ignore the error and return
\code{NULL}. Defaults to "fail".}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 3 columns:
\itemize{
\item \code{arm_num}
\item \code{unique_event_name}
\item \code{form}
}
}
\description{
Execute an "Export Instrument-Event Mappings" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing the mapping between
instruments and events. Note that this request type is not available for
'classic projects', from which event details cannot be exported.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_mapping(conn)
}

}
