% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_repeating.R
\name{meta_repeating}
\alias{meta_repeating}
\title{Fetch repeating instrument names and labels for a REDCap project}
\usage{
meta_repeating(conn, on_error = "fail")
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{on_error}{How to handle errors returned by the API (e.g. events cannot
be exported for classic projects). Set to "fail" to halt execution and
return the API error message, or "null" to ignore the error and return
\code{NULL}. Defaults to "fail".}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 2 columns:
\itemize{
\item \code{instrument_name}
\item \code{instrument_label}
}
}
\description{
Execute an "Export Repeating Instruments and Events" API request to fetch a
\code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing repeating instrument
names and labels.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

meta_repeating(conn)
}

}
