% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xml.R
\name{parse_xml}
\alias{parse_xml}
\title{Convert a REDCap project XML file to a tidy data frame of project records}
\usage{
parse_xml(x)
}
\arguments{
\item{x}{A REDCap project XML obect of class
\code{\link[xml2:oldclass]{xml_document}}, e.g. returned by
\code{\link{project_xml}}}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 6 columns:
\itemize{
\item \code{record_id}
\item \code{form}
\item \code{redcap_event}
\item \code{redcap_repeat_instance}
\item \code{field}
\item \code{value}
}
}
\description{
Extract records from a REDCap project XML file (e.g. returned by
\code{\link{project_xml}}) and assemble into a tidy long-form data frame, with one row
for each combination of record x field x event x instance.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

parse_xml(project_xml(conn))
}

}
