% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_info.R
\name{project_info}
\alias{project_info}
\title{Export REDCap Project information}
\usage{
project_info(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing the columns returned
by an "Export Project Info" API request
}
\description{
Execute an "Export Project Info" API request to fetch project-related details
(e.g. title, creation time, production status, language, etc.) corresponding
to a REDCap project.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

project_info(conn)
}

}
