% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{project_logging}
\alias{project_logging}
\title{Export project logging (audit trail) for a REDCap project}
\usage{
project_logging(
  conn,
  type = NULL,
  user = NULL,
  record = NULL,
  time_start = NULL,
  time_end = NULL
)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{type}{Type of logging to return. Defaults to NULL to return all types.
Specific logging types include:
\itemize{
\item "export": Data export
\item "manage": Manage/Design
\item "user": User or role created-updated-deleted
\item "record": Record created-updated-deleted
\item "record_add": Record created (only)
\item "record_edit": Record updated (only)
\item "record_delete": Record deleted (only)
\item "lock_record": Record locking & e-signatures
\item "page_view": Page Views
}}

\item{user}{REDCap username to fetch logs for. Defaults to NULL to fetch logs
relating to all users. Note that, in the current API version (10.8.5), it's
not possible to pass a character vector with multiple usernames (i.e. one
user per request, or NULL for all).}

\item{record}{Record ID to fetch logs for. Defaults to NULL to fetch logs
relating to all record IDs. Note that, in the current API version (10.8.5),
it's not possible to pass a character vector with multiple record IDs (i.e.
one record per request, or NULL for all).}

\item{time_start}{Fetch logs from \emph{after} a given date/time. Use format
"YYYY-MM-DD HH:MM". Defaults to NULL to omit a lower time limit.}

\item{time_end}{Fetch logs from \emph{before} a given date/time. Use format
"YYYY-MM-DD HH:MM". Defaults to NULL to omit an upper time limit.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 4 columns:
\itemize{
\item \code{timestamp}
\item \code{username}
\item \code{actions}
\item \code{details}
}
}
\description{
Execute an "Export Logging" API request to export the logging (audit trail)
of all changes made to the project, including data exports, data changes,
project metadata changes, modification of user rights, etc.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

project_logging(conn)
}

}
