% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_users.R
\name{project_users}
\alias{project_users}
\title{Export REDCap user information}
\usage{
project_users(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing the columns returned
by an "Export Users" API request
}
\description{
Execute an "Export Users" API request to fetch user-related details (e.g.
username, email, access permissions, etc.) corresponding to a REDCap project.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

project_users(conn)
}

}
