% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_users_dags.R
\name{project_users_dags}
\alias{project_users_dags}
\title{Export REDCap mapping between users and Data Access Groups (DAGs)}
\usage{
project_users_dags(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame with 2 columns:
\itemize{
\item \code{username}
\item \code{redcap_data_access_group}
}
}
\description{
Execute an "Export User-DAG Assignments" API request to fetch the mapping
between users and Data Access Groups (DAGs) associated with a REDCap project.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

project_users_dags(conn)
}

}
