% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_xml.R
\name{project_xml}
\alias{project_xml}
\title{Export REDCap Project XML file}
\usage{
project_xml(
  conn,
  meta_only = FALSE,
  records = NULL,
  fields = NULL,
  events = NULL,
  filter_logic = NULL,
  export_dag = FALSE,
  export_survey = FALSE,
  export_files = FALSE
)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{meta_only}{Logical indicating whether to fetch only the project
metadata (if \code{TRUE}) or both the metadata and data records (if \code{FALSE}).
Defaults to \code{FALSE} to fetch both metadata and data.}

\item{records}{Optional character vector of specific record IDs to fetch
record data for. Only used when \code{meta_only = FALSE}.}

\item{fields}{Optional character vector of specific fields to fetch record
data for. Only used when \code{meta_only = FALSE}.}

\item{events}{Optional character vector of specific events to fetch record
data for. Only used when \code{meta_only = FALSE}.}

\item{filter_logic}{Optional character string containing a REDCap-style
expression used to filter records returned by the API (e.g. "[age] > 30")}

\item{export_dag}{Logical indicating whether to export the
redcap_data_access_group field. Defaults to \code{FALSE}.}

\item{export_survey}{Logical indicating whether to export survey identifier
or timestamp fields, if surveys are used in the project. Defaults to
\code{FALSE}.}

\item{export_files}{Logical indicating whether to export uploaded files. Note
this may lead to large exports. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link[xml2:oldclass]{xml_document}}
}
\description{
Execute an "Export Entire Project as REDCap XML File" API request to fetch
all metadata (and optionally also data records) corresponding to a REDCap
project.
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

project_xml(conn)
}

}
