% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rconn.R
\name{rconn}
\alias{rconn}
\title{Connection to a REDCap Database}
\usage{
rconn(url, token, config = httr::config())
}
\arguments{
\item{url}{URL for a REDCap database API}

\item{token}{REDCap project API token (good practice to set using an
environmental variable, e.g. with \code{\link{Sys.getenv}}).}

\item{config}{Optional configuration settings passed to \code{\link[httr:POST]{httr::POST}}.
Defaults to \code{\link[httr:config]{httr::config()}}.}
}
\value{
An object of class "rconn", to be passed as the first argument to most other
\code{redcap} functions.
}
\description{
Creates an object of class "rconn" containing the URL and token used to
access a REDCap project API.
}
\examples{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

}
