% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_labels.R
\name{recode_labels}
\alias{recode_labels}
\title{Convert between values and labels for factor-type variables (e.g. yes/no,
radio, dropdown, checkbox)}
\usage{
recode_labels(
  x,
  conn,
  dict = redcap::meta_dictionary(conn, add_complete = TRUE),
  convert_to = c("labels", "values"),
  types = c("radio", "yesno", "dropdown", "checkbox"),
  header_labs = FALSE
)
}
\arguments{
\item{x}{Data frame representing a REDCap form (e.g. from a previous export
using \code{\link{fetch_records}})}

\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}

\item{dict}{REDCap metadata dictionary. Defaults to fetching the current
version with \code{\link{meta_dictionary}}}

\item{convert_to}{Convert values to labels ("labels") or labels to values
("values")}

\item{types}{Types of REDCap variables to convert, based on column
"field_type" in the metadata dictionary. Defaults to \code{c("radio", "yesno", "dropdown", "checkbox")}.}

\item{header_labs}{Logical indicating whether column names of \code{x} are labels
(TRUE) or raw variable names (FALSE). Default assumes header has raw
variable names (i.e. FALSE).}
}
\description{
Convert between values and labels for factor-type variables (e.g. yes/no,
radio, dropdown, checkbox)
}
