% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-package.R
\docType{package}
\name{redcap-package}
\alias{redcap}
\alias{redcap-package}
\title{redcap: R Utilities For REDCap}
\description{
R utilities for interacting with the REDCap API.
}
\author{
\strong{Maintainer}: Patrick Barks \email{patrick.barks@epicentre.msf.org}

}
\keyword{internal}
