% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_version.R
\name{redcap_version}
\alias{redcap_version}
\title{Fetch the REDCap database version used for a particular project}
\usage{
redcap_version(conn)
}
\arguments{
\item{conn}{A REDCap API connection object (created with \code{\link{rconn}})}
}
\value{
A character string
}
\description{
Execute an "Export REDCap version" API request
}
\examples{
\dontrun{
conn <- rconn(
  url = "https://redcap.msf.fr/api/",
  token = Sys.getenv("MY_REDCAP_TOKEN")
)

redcap_version(conn)
}

}
